Restructure jobs and add first Job test
