+++ nautobot_circuit_maintenance/jobs.py
@@ -1,0 +1,4 @@
+"""Circuit Maintenance plugin jobs."""
+from .handle_notifications.handler import HandleCircuitMaintenanceNotifications
+
+jobs = [HandleCircuitMaintenanceNotifications]
