--- nautobot_circuit_maintenance/jobs/__init__.py
@@ -1,6 +1,0 @@
-"""Circuit Maintenance jobs."""
-from .handle_notifications import HandleCircuitMaintenanceNotifications
-
-jobs = (HandleCircuitMaintenanceNotifications,)
-
-__all__ = ("jobs",)
